+++ src/main/java/community/solace/spring/cloud/requestreply/service/header/parser/replyindex/MessageReplyIndexParser.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright © Schweizerische Bundesbahnen SBB, 2026.
+ */
+
+package community.solace.spring.cloud.requestreply.service.header.parser.replyindex;
+
+import org.springframework.lang.Nullable;
+import org.springframework.messaging.Message;
+
+@FunctionalInterface
+public interface MessageReplyIndexParser {
+    @Nullable
+    String getReplyIndex(Message<?> message);
+}
+